move lineage context from execution context to its own header file
